+++ package.json
@@ -14,7 +14,7 @@
     "@react-native-community/masked-view": "^0.1.11",
     "@react-native-picker/picker": "1.16.1",
     "@react-navigation/material-bottom-tabs": "^6.0.8",
-    "@react-navigation/native": "^6.0.4",
+    "@react-navigation/native": "^6.0.6",
     "@react-navigation/stack": "^6.0.11",
     "eslint-import-resolver-typescript": "^2.5.0",
     "expo": "~42.0.0",
@@ -39,6 +39,7 @@
     "react-native-switch-selector": "^2.1.4",
     "react-native-vector-icons": "^8.1.0",
     "react-native-web": "^0.17.5",
+    "react-navigation": "^4.4.4",
     "styled-components": "^5.3.3"
   },
   "devDependencies": {
