+++ src/screens/TestScreen.tsx
@@ -1,11 +1,13 @@
 import React from 'react';
+import { Button } from 'react-native';
 import { ScrollPageContainer, InnerPageContainer } from 'screens/styles';
 
-const TestScreen = ({ navigation }: any) => {
+const TestScreen = ({ navigation: { goBack } }: any) => {
   return (
     <ScrollPageContainer>
       <InnerPageContainer>
         <p>Hello</p>
+        <Button title="Go Back" onPress={() => goBack()} />
       </InnerPageContainer>
     </ScrollPageContainer>
   );
